turnos integrados en la agenda

+++ apps/gestion-casa/entity/turnos/controller.ts
@@ -1,52 +1,44 @@
 import {  StatusCodes } from "../../../../dep/deps.ts";
-import  { Prisma } from "../../generated/client/deno/edge.ts";
 import   prisma  from "../../prisma/db.ts";
 
 
-
-const get= async (ctx: any) => {
-
-    const data = await prisma.turno.findMany();
-    ctx.response.status = 201;
-    ctx.response.body = {
-      status: StatusCodes.OK,
-      data,
-    };
-};
-
-
-const add = async (ctx: any) => {
+  const AllOperations = async (ctx: any) =>  {
     try {
-      const turno: Prisma.turnoCreateInput = await ctx.request.body().value;
+  
+      const {fecha, quien} = await ctx.request.body().value;
+
+      const turno = await prisma.turno.findFirst({where : {fecha}});
+      const id= turno?.id;
+  
+      if(quien){
+        if(turno){
+          //update
+          if(turno.quien!= quien){
+          await prisma.turno.updateMany({
+              where: {id },
+              data: {quien}
+            })
+          }
+  
+        }
+        else{
+          await prisma.turno.create({
+            data: {fecha, quien}
+          });
+        }
+  
+      }
+      else{
+        if(turno){
+          await prisma.turno.deleteMany({where: {id}});
+        }
+      }
       
-      const data = await prisma.turno.create({
-        data: turno
-      });
-  
-  
-      ctx.response.status = 201;
-      ctx.response.body = {
-        status: StatusCodes.OK,
-        data,
-      };
-    } catch (error) {
-      ctx.response.status = 500;
-      ctx.response.body = { status: StatusCodes.INTERNAL_SERVER_ERROR, message: error.message };
-      return;
-    }
-  };
-
-  
-const del = async (ctx: any) =>  {
-    try {
-  
-      const  id  = Number(ctx?.params?.id);
-      const data = await prisma.turno.deleteMany({where: {id}});
   
       ctx.response.status = 200;
       ctx.response.body = {
         status: StatusCodes.OK,
-        data,
+        data : {ok : 'yes'},
       };
     } catch (error) {
       ctx.response.status = 500;
@@ -56,11 +48,13 @@ const del = async (ctx: any) =>  {
       };
       return;
     }
-  };
+
+
+    
+
+  }
 
 
 export default { 
-    get,
-    add, 
-    del
+    AllOperations
 };
